Data-moved Entites into Data, added organizational foldesr, renamed context, deleted old Entities project.
