+++ PicAFlick.Domain.Services/PicAFlick.Domain.Services.csproj
@@ -9,7 +9,6 @@
   <ItemGroup>
     <ProjectReference Include="..\PicAFlick.Domain.Core\PicAFlick.Domain.Core.csproj" />
     <ProjectReference Include="..\PicAFlick.ExternalDtos\PicAFlick.ExternalDtos.csproj" />
-    <ProjectReference Include="..\PicAFlick.Domain.Entities\PicAFlick.Domain.Entities.csproj" />
 	<ProjectReference Include="..\PicAFlick.Data\PicAFlick.Data.csproj" />
   </ItemGroup>
