+++ PicAFlick.WebApi/PicAFlick.WebApi.csproj
@@ -18,7 +18,6 @@
     <ProjectReference Include="..\PicAFlick.Data\PicAFlick.Data.csproj" />
     <ProjectReference Include="..\PicAFlick.Domain.Core\PicAFlick.Domain.Core.csproj" />
     <ProjectReference Include="..\PicAFlick.ExternalDtos\PicAFlick.ExternalDtos.csproj" />
-    <ProjectReference Include="..\PicAFlick.Domain.Entities\PicAFlick.Domain.Entities.csproj" />
     <ProjectReference Include="..\PicAFlick.Domain.Services\PicAFlick.Domain.Services.csproj" />
   </ItemGroup>
