+++ PicAFlick.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.12.35707.178
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PicAFlick.Domain.Entities", "PicAFlick.Domain.Entities\PicAFlick.Domain.Entities.csproj", "{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PicAFlick.Domain.Services", "PicAFlick.Domain.Services\PicAFlick.Domain.Services.csproj", "{9AB456EB-3F43-4110-A869-DA292B775D15}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PicAFlick.Data", "PicAFlick.Data\PicAFlick.Data.csproj", "{2EBFEB1C-6C1A-4027-AAD7-2B485F67FECB}"
@@ -32,18 +30,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|x64.Build.0 = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Debug|x86.Build.0 = Debug|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|x64.ActiveCfg = Release|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|x64.Build.0 = Release|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|x86.ActiveCfg = Release|Any CPU
-		{DA7E6952-5C1C-43A6-BBA6-EC4F6D72BAD5}.Release|x86.Build.0 = Release|Any CPU
 		{9AB456EB-3F43-4110-A869-DA292B775D15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9AB456EB-3F43-4110-A869-DA292B775D15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9AB456EB-3F43-4110-A869-DA292B775D15}.Debug|x64.ActiveCfg = Debug|Any CPU
